Move `SkinnableInfo` to better namespace
