+++ osu.Game/Extensions/DrawableExtensions.cs
@@ -5,7 +5,6 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Configuration;
-using osu.Game.Screens.Play.HUD;
 using osu.Game.Skinning;
 using osuTK;
