+++ osu.Game/Skinning/ISkinnableTarget.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Game.Extensions;
-using osu.Game.Screens.Play.HUD;
+using osu.Game.Rulesets;
 
 namespace osu.Game.Skinning
 {
