+++ osu.Game/Skinning/SkinnableInfo.cs
@@ -13,10 +13,10 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Logging;
 using osu.Game.Configuration;
 using osu.Game.Extensions;
-using osu.Game.Skinning;
+using osu.Game.Rulesets;
 using osuTK;
 
-namespace osu.Game.Screens.Play.HUD
+namespace osu.Game.Skinning
 {
     /// <summary>
     /// Serialised information governing custom changes to an <see cref="ISkinnableDrawable"/>.
