+++ osu.Game/Skinning/SkinnableTargetContainer.cs
@@ -8,7 +8,7 @@ using System.Threading;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Screens.Play.HUD;
+using osu.Game.Skinning.Serialisation;
 
 namespace osu.Game.Skinning
 {
